+++ src/js/language/it_IT.js
@@ -361,6 +361,7 @@
                     FR: 'Francia',
                     GB: 'Regno Unito',
                     IE: 'Irlanda',
+                    IN: 'India',
                     IT: 'Italia',
                     MA: 'Marocco',
                     NL: 'Paesi Bassi',