+++ src/js/language/no_NO.js
@@ -361,6 +361,7 @@
                     FR: 'Frankrike',
                     GB: 'Storbritannia',
                     IE: 'Irland',
+                    IN: 'India',
                     IT: 'Italia',
                     MA: 'Marokko',
                     NL: 'Nederland',